requirements.txt
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
beautifulsoup4==4.12.2
schedule==1.2.0
gitpython==3.1.40
python-dotenv==1.0.0
psutil==5.9.6
selenium==4.15.2